+++ my-app/about.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html>
+<html lang="fr">
 
 <head>
 	<meta charset="UTF-8" />
@@ -10,7 +10,7 @@
 	<link rel="stylesheet" media="screen" type="text/css" href="./css/responsive.css" />
 	<script src="https://code.iconify.design/3/3.1.0/iconify.min.js"></script>
 	<title>EnviEndoc - About</title>
-	<script type="module" src="./js/pageTitles.js"></script>
+	<script type="module" src="./public/js/pageTitles.js"></script>
 
 </head>
 
@@ -26,7 +26,7 @@
             </ul>
         </nav>
         <div class="box">
-            <select id="language-select" onchange="selectLang()">
+            <select id="language-select">
                 <option value="fr">Français</option>
                 <option value="en">English</option>
             </select>
@@ -285,8 +285,8 @@
 			</nav>
 		</div>
 	</footer>
-	<script type="module" src="./js/copyright.js"></script>
-	<script type="module" src="./js/translate.js"></script>
+	<script type="module" src="./public/js/copyright.js"></script>
+	<script type="module" src="./public/js/translate.js"></script>
 </body>
 
 </html>
