+++ my-app/css/style.css
@@ -1,5 +1,6 @@
 @import "../node_modules/ol/ol.css";
 
+
 /* Déclarations de variable pour uniformiser les pages,
  couleurs, polices etc
  changer couleurs avec les exactes*/
