--- my-app/function.js
@@ -1,3 +1,0 @@
-function NavigationAccueil(){
-    navigate('http://localhost:5173/#');
-}
