+++ my-app/list.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html>
+<html lang="fr">
 
 <head>
 	<meta charset="UTF-8" />
@@ -9,7 +9,7 @@
 	<link rel="stylesheet" media="screen" type="text/css" href="./css/style.css" />
 	<link rel="stylesheet" media="screen" type="text/css" href="./css/responsive.css" />
 	<title>EnviEndoc - List of endocrine disruptors</title>
-	<script type="module" src="./js/pageTitles.js"></script>
+	<script type="module" src="./public/js/pageTitles.js"></script>
 </head>
 
 <body>
@@ -24,7 +24,7 @@
             </ul>
         </nav>
         <div class="box">
-            <select id="language-select" onchange="selectLang()">
+            <select id="language-select">
                 <option value="fr">Français</option>
                 <option value="en">English</option>
             </select>
@@ -82,9 +82,9 @@
 			</nav>
 		</div>
 	</footer>
-	<script type="module" src="./js/copyright.js"></script>
-	<script type="module" src="./js/readcsv.js"></script>
-	<script type="module" src="./js/translate.js"></script>
+	<script type="module" src="./public/js/copyright.js"></script>
+	<script type="module" src="./public/js/readcsv.js"></script>
+	<script type="module" src="./public/js/translate.js"></script>
 </body>
 
 </html>
