+++ my-app/public/images/Equipe/SylvainChamot.jpg
