+++ my-app/public/js/copyright.js
